+++ crates/utils/sov-proxy-utils/src/node_check_metric.rs
@@ -1,11 +1,33 @@
-use crate::root_hash_checker::RootHashCheck;
-use crate::root_hash_checker::RootHashConsistency;
+use crate::node_checker::RootHashCheck;
+use crate::LatestHeightCheckStats;
 use std::collections::BTreeSet;
 use std::io::Write;
 
 #[derive(Debug)]
+pub struct LatestHeightCheckMetric {
+    pub stats: LatestHeightCheckStats,
+}
+
+impl sov_metrics::Metric for LatestHeightCheckMetric {
+    fn measurement_name(&self) -> &'static str {
+        "sov_proxy_latest_height_check"
+    }
+
+    fn serialize_for_telegraf(&self, buffer: &mut Vec<u8>) -> std::io::Result<()> {
+        let stats = &self.stats;
+        write!(
+            buffer,
+            "{} nodes_ok={},nodes_failed={},height_diff={}",
+            self.measurement_name(),
+            stats.nodes_ok,
+            stats.nodes_failed,
+            stats.height_diff,
+        )
+    }
+}
+
+#[derive(Debug)]
 pub struct RootHashCheckMetric {
-    pub outcome: RootHashConsistency,
     pub slot_number: u64,
     pub nodes_ok: u64,
     pub nodes_failed: u64,
@@ -13,25 +35,16 @@ pub struct RootHashCheckMetric {
 }
 
 impl RootHashCheckMetric {
-    pub fn from_check(check: &RootHashCheck, outcome: RootHashConsistency) -> Self {
+    pub(crate) fn from_check(check: &RootHashCheck) -> Self {
         let nodes_ok = check.node_results.len() as u64;
         let nodes_failed = check.failed_nodes.len() as u64;
         let unique_state_roots = check.node_results.values().collect::<BTreeSet<_>>().len() as u64;
 
         Self {
-            outcome,
             slot_number: check.slot_number,
             nodes_ok,
             nodes_failed,
             unique_state_roots,
-        }
-    }
-
-    fn outcome_tag(&self) -> &'static str {
-        match self.outcome {
-            RootHashConsistency::AllMatch => "all_match",
-            RootHashConsistency::Mismatch => "mismatch",
-            RootHashConsistency::NoData => "no_data",
         }
     }
 }
@@ -44,9 +57,8 @@ impl sov_metrics::Metric for RootHashCheckMetric {
     fn serialize_for_telegraf(&self, buffer: &mut Vec<u8>) -> std::io::Result<()> {
         write!(
             buffer,
-            "{},outcome={} slot_number={},nodes_ok={},nodes_failed={},unique_state_roots={}",
+            "{} slot_number={},nodes_ok={},nodes_failed={},unique_state_roots={}",
             self.measurement_name(),
-            self.outcome_tag(),
             self.slot_number,
             self.nodes_ok,
             self.nodes_failed,